+++ R/appraise.R
@@ -3,8 +3,11 @@
 #' @param model a fitted model. Currently models inheriting from class `"gam"`,
 #'   as well as classes `"glm"` and `"lm"` from calls to [stats::glm] or
 #'   [stats::lm] are supported.
-#' @param method character; method used to generate theoretical quantiles.
-#'   The default is `"uniform"`, which generates reference quantiles using
+#' @param method character or `NULL`; method used to generate theoretical
+#'   quantiles. The default, `NULL`, selects `"simulate"` for models fitted with
+#'   [mgcv::tw()] or [mgcv::Tweedie()] and `"uniform"` otherwise. Explicitly
+#'   supplying a method overrides this selection, subject to availability of
+#'   the required family functions. `"uniform"` generates reference quantiles using
 #'   random draws from a uniform distribution and the inverse cumulative
 #'   distribution function (CDF) of the fitted values. The reference quantiles
 #'   are averaged over `n_uniform` draws. `"simulate"` generates reference
@@ -12,10 +15,11 @@
 #'   values of the covariates, which are then residualised to generate reference
 #'   quantiles, using `n_simulate` simulated data sets. `"normal"` generates
 #'   reference quantiles using the standard normal distribution. `"uniform"` is
-#'   more computationally efficient, but `"simulate"` allows reference bands to
+#'   often more computationally efficient, but Tweedie quantiles are expensive
+#'   to compute. `"simulate"` allows reference bands to
 #'   be drawn on the QQ-plot. `"normal"` should be avoided but is used as a fall
 #'   back if a random number generator (`"simulate"`) or the inverse of the CDF
-#'   (`"uniform"``) are not available from the `family` used during model
+#'   (`"uniform"`) are not available from the `family` used during model
 #'   fitting.
 #'
 #'   Note that `method = "direct"` is deprecated in favour of
@@ -88,7 +92,7 @@
 #' @export
 `appraise.gam` <- function(
   model,
-  method = c("uniform", "simulate", "normal", "direct"),
+  method = NULL,
   use_worm = FALSE,
   n_uniform = 10,
   n_simulate = 50,
@@ -107,11 +111,6 @@
   ...
 ) {
   ## process args
-  method <- match.arg(method)
-  if (identical(method, "direct")) {
-    message("`method = \"direct\"` is deprecated, use `\"uniform\"`")
-    method <- "uniform"
-  }
   type <- match.arg(type)
   if (is.character(n_bins)) {
     n_bins <- match.arg(n_bins)